Simple refactoring of explorer pools.
