+++ ts/a11y/explorer.ts
@@ -31,11 +31,8 @@ import {OptionList, expandable} from '../util/Options.js';
 import {SerializedMmlVisitor} from '../core/MmlTree/SerializedMmlVisitor.js';
 import {MJContextMenu} from '../ui/menu/MJContextMenu.js';
 
-import {Explorer} from './explorer/Explorer.js';
 import * as ke from './explorer/KeyExplorer.js';
-import * as me from './explorer/MouseExplorer.js';
-import {TreeColorer, FlameColorer} from './explorer/TreeExplorer.js';
-import {LiveRegion, SpeechRegion, ToolTip, HoverRegion} from './explorer/Region.js';
+import { ExplorerPool, RegionPool } from './explorer/ExplorerPool.js';
 
 import {Submenu} from 'mj-context-menu/js/item_submenu.js';
 
@@ -97,7 +94,7 @@ export function ExplorerMathItemMixin<B extends Constructor<HTMLMATHITEM>>(
     /**
      * The Explorer objects for this math item
      */
-    protected explorers: {[key: string]: Explorer} = {};
+    protected explorers: ExplorerPool;
 
     /**
      * The currently attached explorers
@@ -135,7 +132,7 @@ export function ExplorerMathItemMixin<B extends Constructor<HTMLMATHITEM>>(
           this.savedId = null;
         }
         // Init explorers:
-        this.explorers = initExplorers(document, node, mml);
+        this.explorers = new ExplorerPool(document, node, mml);
         this.attachExplorers(document);
       }
       this.state(STATE.EXPLORER);
@@ -149,8 +146,8 @@ export function ExplorerMathItemMixin<B extends Constructor<HTMLMATHITEM>>(
     public attachExplorers(document: ExplorerMathDocument) {
       this.attached = [];
       let keyExplorers = [];
-      for (let key of Object.keys(this.explorers)) {
-        let explorer = this.explorers[key];
+      for (let key of Object.keys(this.explorers.explorers)) {
+        let explorer = this.explorers.explorers[key];
         if (explorer instanceof ke.AbstractKeyExplorer) {
           explorer.AddEvents();
           explorer.stoppable = false;
@@ -180,7 +177,7 @@ export function ExplorerMathItemMixin<B extends Constructor<HTMLMATHITEM>>(
       this.savedId = this.typesetRoot.getAttribute('sre-explorer-id');
       this.refocus = (window.document.activeElement === this.typesetRoot);
       for (let key of this.attached) {
-        let explorer = this.explorers[key];
+        let explorer = this.explorers.explorers[key];
         if (explorer.active) {
           this.restart.push(key);
           explorer.Stop();
@@ -195,7 +192,7 @@ export function ExplorerMathItemMixin<B extends Constructor<HTMLMATHITEM>>(
     public updateDocument(document: ExplorerMathDocument) {
       super.updateDocument(document);
       this.refocus && this.typesetRoot.focus();
-      this.restart.forEach(x => this.explorers[x].Start());
+      this.restart.forEach(x => this.explorers.explorers[x].Start());
       this.restart = [];
       this.refocus = false;
     }
@@ -212,7 +209,7 @@ export interface ExplorerMathDocument extends HTMLDOCUMENT {
   /**
    * The objects needed for the explorer
    */
-  explorerRegions: ExplorerRegions;
+  explorerRegions: RegionPool;
 
   /**
    * Add the Explorer to the MathItems in the MathDocument
@@ -276,7 +273,7 @@ export function ExplorerMathDocumentMixin<B extends MathDocumentConstructor<HTML
     /**
      * The objects needed for the explorer
      */
-    public explorerRegions: ExplorerRegions;
+    public explorerRegions: RegionPool = new RegionPool(this);
 
     /**
      * Extend the MathItem class used for this MathDocument
@@ -294,8 +291,6 @@ export function ExplorerMathDocumentMixin<B extends MathDocumentConstructor<HTML
       const visitor = new SerializedMmlVisitor(this.mmlFactory);
       const toMathML = ((node: MmlNode) => visitor.visitTree(node));
       this.options.MathItem = ExplorerMathItemMixin(this.options.MathItem, toMathML);
-      // TODO: set backward compatibility options here.
-      this.explorerRegions = initExplorerRegions(this);
     }
 
     /**
@@ -350,115 +345,6 @@ export function ExplorerHandler(handler: HANDLER, MmlJax: MATHML = null): HANDLE
 
 
 /*==========================================================================*/
-
-/**
- * The regions objects needed for the explorers.
- */
-export type ExplorerRegions = {
-  speechRegion?: SpeechRegion,
-  brailleRegion?: LiveRegion,
-  magnifier?: HoverRegion,
-  tooltip1?: ToolTip,
-  tooltip2?: ToolTip,
-  tooltip3?: ToolTip
-};
-
-
-/**
- * Initializes the regions needed for a document.
- * @param {ExplorerMathDocument} document The current document.
- */
-function initExplorerRegions(document: ExplorerMathDocument) {
-  return {
-    speechRegion: new SpeechRegion(document),
-    brailleRegion: new LiveRegion(document),
-    magnifier: new HoverRegion(document),
-    tooltip1: new ToolTip(document),
-    tooltip2: new ToolTip(document),
-    tooltip3: new ToolTip(document)
-  };
-}
-
-
-
-/**
- * Type of explorer initialization methods.
- * @type {(ExplorerMathDocument, HTMLElement, any[]): Explorer}
- */
-type ExplorerInit = (doc: ExplorerMathDocument,
-                     node: HTMLElement, ...rest: any[]) => Explorer;
-
-/**
- *  Generation methods for all MathJax explorers available via option settings.
- */
-let allExplorers: {[options: string]: ExplorerInit} = {
-  speech: (doc: ExplorerMathDocument, node: HTMLElement, ...rest: any[]) => {
-    let explorer = ke.SpeechExplorer.create(
-      doc, doc.explorerRegions.speechRegion, node, ...rest) as ke.SpeechExplorer;
-    explorer.speechGenerator.setOptions({
-      automark: true as any, markup: 'ssml_step',
-      locale: doc.options.sre.locale, domain: doc.options.sre.domain,
-      style: doc.options.sre.style, modality: 'speech'});
-    // This weeds out the case of providing a non-existent locale option.
-    let locale = explorer.speechGenerator.getOptions().locale;
-    if (locale !== Sre.engineSetup().locale) {
-      doc.options.sre.locale = Sre.engineSetup().locale;
-      explorer.speechGenerator.setOptions({locale: doc.options.sre.locale});
-    }
-    explorer.showRegion = 'subtitles';
-    return explorer;
-  },
-  braille: (doc: ExplorerMathDocument, node: HTMLElement, ...rest: any[]) => {
-    let explorer = ke.SpeechExplorer.create(
-      doc, doc.explorerRegions.brailleRegion, node, ...rest) as ke.SpeechExplorer;
-    explorer.speechGenerator.setOptions({automark: false as any, markup: 'none',
-                                         locale: 'nemeth', domain: 'default',
-                                         style: 'default', modality: 'braille'});
-    explorer.showRegion = 'viewBraille';
-    return explorer;
-  },
-  keyMagnifier: (doc: ExplorerMathDocument, node: HTMLElement, ...rest: any[]) =>
-    ke.Magnifier.create(doc, doc.explorerRegions.magnifier, node, ...rest),
-  mouseMagnifier: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    me.ContentHoverer.create(doc, doc.explorerRegions.magnifier, node,
-                             (x: HTMLElement) => x.hasAttribute('data-semantic-type'),
-                             (x: HTMLElement) => x),
-  hover: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    me.FlameHoverer.create(doc, null, node),
-  infoType: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    me.ValueHoverer.create(doc, doc.explorerRegions.tooltip1, node,
-                           (x: HTMLElement) => x.hasAttribute('data-semantic-type'),
-                           (x: HTMLElement) => x.getAttribute('data-semantic-type')),
-  infoRole: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    me.ValueHoverer.create(doc, doc.explorerRegions.tooltip2, node,
-                           (x: HTMLElement) => x.hasAttribute('data-semantic-role'),
-                           (x: HTMLElement) => x.getAttribute('data-semantic-role')),
-  infoPrefix: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    me.ValueHoverer.create(doc, doc.explorerRegions.tooltip3, node,
-                           (x: HTMLElement) => x.hasAttribute('data-semantic-prefix'),
-                           (x: HTMLElement) => x.getAttribute('data-semantic-prefix')),
-  flame: (doc: ExplorerMathDocument, node: HTMLElement, ..._rest: any[]) =>
-    FlameColorer.create(doc, null, node),
-  treeColoring: (doc: ExplorerMathDocument, node: HTMLElement, ...rest: any[]) =>
-    TreeColorer.create(doc, null, node, ...rest)
-};
-
-
-/**
- * Initialises explorers for a document.
- * @param {ExplorerMathDocument} document The target document.
- * @param {HTMLElement} node The node explorers will be attached to.
- * @param {string} mml The corresponding Mathml node as a string.
- * @return {Explorer[]} A list of initialised explorers.
- */
-function initExplorers(document: ExplorerMathDocument, node: HTMLElement, mml: string): {[key: string]: Explorer} {
-  let explorers: {[key: string]: Explorer} = {};
-  for (let key of Object.keys(allExplorers)) {
-    explorers[key] = allExplorers[key](document, node, mml);
-  }
-  return explorers;
-}
-
 
 /* Context Menu Interactions */
